gibrc: fix less, improve pathadd
Getting less to work properly is an absolute pain. Here's the best
solution I can find.

Adds less as a binary in ~/bin. The binary checks whether less has a
file argument, or is reading from stdin.

If it's < one screen and a file, less it with -FX.
If it's > one screen and a file, use normal less.

If it's from stdin just cat it to the stdout until it's more than one
screen, at which point run it all through normal less.

I don't know of another way to get the -FX style "cat if less than one
screen", while also preserving ability to scroll (which -X removes).

Refs: https://unix.stackexchange.com/q/167735/144642