helpers: add scripts to print 256 terminal colours
When testing colours a single-page way to see all the colours you have
to choose from is really helpful, especially when it shows you what
things look like in your terminal.

This adds two scripts, one which prints a block of colour, and one which
prints text in the colour.

Refs: https://askubuntu.com/a/821163/194570