gibrc: use zsh/bash on sshed systems if possible
If we ssh into a system we want to use zsh or bash if possible, and
we might not have admin (or want to `chsh -s` for our account). This
defaults the shell to zsh or bash (if available).

Note that if you actively run a different shell after logging in, that
will still work (e.g. you can run `bash` to spawn a bash subshell inside
your zsh shell).

This also doesn't affect scripts due to the interactive check at the
beginning of gibrc.