gibrc: upgrade zsh theme to be async
Fairly major change that reworks the prompt. This is based on https://github.com/sindresorhus/pure, but changed to suit me.

This makes zsh actually usable on AIX or on slow remote machines for
example.

This also updates the bash theme to make it match the zsh one, while
still hopefully looking different enough that you know you're not in
zsh.