karabiner: fix Ж-d date generation
Karabiner doesn't wait for the first to action to finish before running
the second one, so what is pasted is the output of the previous copy
command.

Fix this by running the copy command when the key is pressed, and the
paste when the key is released. The computer should be faster than the
hand (especially once the signal has been debounced) so this should
work.