gibrc: vulkan sdk and fzf updates
Adds a "preview" command (preview list of files, open them in $VISUAL
with <Ctrl-o>).

Adds the vulkan sdk (installed with brew on Darwin, haven't worked
whether it's needed for Linux yet).

Also sets the TERM field when using the `s` ssh alias, so ssh'ing to
terminals works when you're not using `xterm-256color`.