git: add lp (log preview) command using fzf
Not sure whether it's better to have this as a git alias or keybindings,
`g lp` is easy to remember, but if you want to checkout a commit hash,
then typing:

```bash
g co `g lp`
```

Is more annoying than just typing `g co <c-g><c-l>` (or equivalent).

Either way, this replaces `tig` for me entirely, which is nice.

Refs: https://junegunn.kr/2016/07/fzf-git