less: go back to vanilla less
Hooray! The patch to fix -RF in less without requiring -X and breaking
scrolling is in newer versions of less (fixed in version 520, broken in
version 487). So we can get rid of the shell script.

Also adds some other niceties as less options, including colors in
manpages (try `man less`).

Looks like Ubuntu still isn't on a newer version of less with the fix,
so will need to update to a newer less soon.

Refs: https://www.topbug.net/blog/2016/09/27/make-gnu-less-more-powerful/
Refs: http://www.greenwoodsoftware.com/less/news.520.html
Refs: https://github.com/sharkdp/bat/blob/80da0dc6194d94dbbc87185283d4e299b8101dfd/README.md#using-a-different-pager
Refs: https://packages.ubuntu.com/search?keywords=less