vimrc,gibrc: remove kj -> Esc mappings
Use <kbd>Esc</kbd><kbd>Esc</kbd> to go to normal mode in terminal mode,
and a single <kbd>Esc</kbd> to go to normal mode in the shell (inside
vim).

Answer to the <kbd>Esc</kbd> key being too far away is to map it nearer,
having to press two keys every time is a pain, and it makes typing
commands that end in `k` annoying.