vimrc: use echodoc plugin, always show gutter
The echodoc plugin shows you the function signature of what you're
typing as you type it. Using experimental APIs in nvim 0.3.2 (currently
in prerelease) you can display it next to the cursor.

Always show the gutter so it doesn't keep popping in and out.

Increase the memory given to the java language server, hopefully this
will make it faster.