macOS: fix setup shorcut defaults checking with \ inputs
Sometimes `defaults read` doubles the number of backslashes it shows,
leading to a situation where every time you run the update script it
doubles the number of backslashes in the shorcut string.

Fixes compatibility with /usr/bin/sed on macOS, as the builtin sed has
no support for character classes like `\s` or `\W`.