vimrc, zsh: improve plugin handling, reinstate kj
- Automatically install vim-plug if it isn't already installed [0][].
- Have vim check the nvim autoload directory (avoids symlinks).
- Always load all the plugins, just don't enable vim ones on nvim (and
  vice-versa). This means `:PlugClean` doesn't try to remove them [1][].
- Make <Tab> toggle the completion menu or indent smartly
  (:h deoplete#manual_complete())
- Move the vim-surround shortcuts operator-pending shortcuts so neovim
  shows the _ cursor while the operator is pending [2][].
- Slightly improve printing in the link script.
- Improve the shortcut to quickly close the LanguageClient-neovim preview
  windows [3][].
- Add shortcuts to close the windows around you with d<C-h/i/j/k> [3][].

---

This also tries to come to an optimal solution for the <Esc> problem in
zsh and vim. Basically when you run a terminal inside vim or nvim,
<kbd>Esc</kbd> can have to meanings:
  1. Go to zsh normal mode
  2. Go to (n)vim normal mode

My solution is to use <kbd>Esc</kbd> for (n)vim normal mode, and to
reinstate kj for zsh normal mode. Outside (n)vim, <kbd>Esc</kbd> still
works for zsh normal mode. This is ideal because kj in zsh adds only a
0.1s delay if you hit k and then wait, but in (n)vim it takes longer. In
general it makes more sense to remap <kbd>Esc</kbd> to be closer, than
to use `kj`, as <kbd>Esc</kbd> is still a single key.

[0]: https://github.com/junegunn/vim-plug/wiki/tips#automatic-installation
[1]: https://github.com/junegunn/vim-plug/wiki/tips#conditional-activation
[2]: https://vi.stackexchange.com/questions/16565/reclaim-cursor-shape-in-operator-pending-mode-neovim-while-using-vim-surround
[3]: https://github.com/autozimu/LanguageClient-neovim/issues/645