gibrc: fix mvm and cpm commands
`cp` and `mv` do two different things depending on whether the last arg
ends with a `/`. Instead of doing the wrong thing when there isn't a
trailing slash, do the right thing when there isn't, and error when
there is.

Also increases kitty scrollback to a giant value, doesn't seem to cause
instability.