gibrc: fix first run in zsh, add clc alias
Previously when you first set up a machine, the
`"$XDG_DATA_HOME"/zfunc/source/*` glob matches nothing as there are no
source files yet. Set the `nullglob` option to ignore this error, then
unset it if we set it to avoid changing it permanently.

Add `clc` to copy the last command run.