vim: revert Colemak remappings
Having played with this for a couple of days, it just doesn't seem worth
the cognitive load it adds, especially as there are lots of vim commands
that one uses in other apps (e.g. `n` to go to the next match in less).

Having gotten used to using the arrow keys for arrow movements, it's not
a huge amount more painful. For most movements I should be
de-emphasizing the arrows vs other movements anyway.