kitty: update config, change history pager direction
This uses `/` instead of `?` with less to make n/N work as you'd expect
them to.

Also removes the status column so things aren't all wrapped one or more
columns, so the display looks like it did in the terminal.

This does mean that you won't always scroll to the last prompt, but
that's usually clearer.

It does also mean that if there are no prompt characters on the current
screen (i.e. you press ⌘-f while running a long running command with
more than one screen of output) you'll have to press N again to get to
the previous prompt, but that's worth it for having the shortcuts the
right way round.