feat(vim): add commands to generate XXX(gib) comments
Also fail the commit for the default pipeline if XXX comments are found.

The rationale is that XXX comments should be fixed immediately, and
should never be committed or pushed anywhere (they're more notes to self
to come back to in the course of making changes).

TODO comments by comparison are for things that should be fixed later,
possibly by other people.