refactor(mac): pull out defaults reading code into a function
This allows us to handle the chaos of `defaults read` on a bool
returning an integer value.

Also by echoing to stdout when the value is changed, we can trigger any
commands needed to make macOS reread the value.

As part of this I had to change some of the ints to bools to match what
was already in the defaults.