fix(zsh): don't show the job creation output
Normally when you run a background job you get:

```
[2] 1871
```

when the job is kicked off, and:

```
[2]  + 1871 done       git fetch --all --quiet
```

when the job terminates.

I want the last one to know that the job completed, but I don't need the
first one, so set no_monitor only within this function to get the
desired behaviour.