fix(macOS): handle CopyQ accessibility preferences
Whenever CopyQ gets updated it loses its Accessibility permissions,
despite still being ticked in System Preferences. This is because the
app is unsigned, so when it is given accessibility permissions, that
is only given for the requested binary. When a new binary is installed
(the app is updated) even though it shows as ticked, it has in fact not
got access.