fix(hammerspoon): go back to real mute/unmute shortcuts, handle Webex
Webex recognises that you've muted the microphone globally, and mutes
itself too. However when you unmute the microphone globally, Webex
doesn't notice, so you stay muted forever.

Work around this by doing the equivalent of Cmd-Tabbing to Webex and
clicking Participant > Unmute Me.

This solves the unmuting problem, but there's another issue. If you mute
and your mic volume is set to 100%, when you unmute Webex sets it to
aroudn 30%, even if you untick "Automatically adjust volume". Manually
setting it back up to 100% doesn't seem to work either, probably because
Webex hasn't processed the unmuting fully when I change the volume back.

What seems to work for me is hitting the unmute hotkey twice. Sad but it
seems to work.