feat(vim): switch from vim-sneak to lightspeed.vim
After some interesting reading on the latest in the sneak plugin space
(especially since a whole new crop of lua-based plugins appeared for
Neovim), ended up finding lightspeed to be the best fit for the
movements I'm looking for.

Lightspeed doesn't support case-insensitive searching, which will
require a little mental work, but overall it should be very worth it.

Refs: https://medium.com/codex/vim-for-the-win-the-power-of-single-char-motion-82d93b4bf6d6