fix(nvim): get noice/nvim-notify notifications out of the way
Often I'll load up nvim, especially in a half-width tab when on my
built-in laptop screen, and large notifications will completely block
the thing I'm looking at. Common offenders are crashing Java
languageservers with stacktraces.

Ideally nvim-notify would be able to keep out of the way of the cursor
so I could always see what I was doing, but in the absence of that, at
least starting from the bottom right vs the top right makes overlap less
likely.