feat(git): add `git-sw` command that can handle any branch switching
Don't want to have to think about whether it's a new branch, or a
commit, or something else. I just want the `switch` flags to work, and
the `upstream` to be configured to what makes sense.

This breaks tab-completion (I don't know how to get `git sw` to
tab-complete as `git switch`), so modify Ctrl-g Ctrl-b to output the
branch directly. For commit hashes Ctrl-g Ctrl-h already works.

Might need some tweaking for further edge-cases.