fix(nvim): use Tab for completion
Shift-Enter is a bit hard to reach for something one does as often as
accepting the current completion. Changing between options is easier
with the arrow keys anyway (on a keyboard with arrow keys on the home
row), so we can use Tab for completion. With some fanciness we can also
continue to use it to jump between snippets.